+++ app/controllers/member/sharings_controller.rb
@@ -43,7 +43,7 @@ class Member::SharingsController < Member::BaseController
 
   def destroy
     @sharing = @group.sharings.find params[:id]
-    if @sharing && (@sharing.shared_by == current_user || @group.moderators.include?(current_user) )
+    if @sharing && (@sharing.user == current_user || @group.moderators.include?(current_user) )
       @sharing.destroy         
       flash[:ok] = I18n.t("tog_social.sharings.member.remove_share_ok", :name => @group.name)
     else