+++ app/models/user.rb
@@ -16,7 +16,7 @@ class User < ActiveRecord::Base
   has_many :moderated_groups, :through => :moderator_memberships,
                     :conditions => "memberships.state='active' and groups.state='active'", :source => :group
 
-  has_many :sharings, :class_name => 'GroupSharing', :dependent => :destroy, :foreign_key => 'shared_by'
+  has_many :sharings, :class_name => 'Share', :dependent => :destroy
   
   # => oauth support
   has_many :client_applications