+++ init.rb
@@ -1,8 +1,8 @@
 require_plugin 'tog_core'
-require_plugin 'tog_mail'
 require_plugin 'acts_as_state_machine'
 require_plugin 'seo_urls'
 require_plugin 'paperclip'
+require_plugin 'acts_as_shareable'
 
 require "i18n" unless defined?(I18n)
 Dir[File.dirname(__FILE__) + '/locale/**/*.yml'].each do |file|